fixes #24242 - pulpnode for storage notifications

+++ app/services/katello/ui_notifications/pulp/proxy_disk_space.rb
@@ -5,7 +5,9 @@ module Katello
         class << self
           def deliver!
             SmartProxy.unscoped.with_content.each do |proxy|
-              percentage = proxy.statuses[:pulp].storage['pulp_dir']['percent']
+              status = proxy.statuses[:pulp] || proxy.statuses[:pulpnode]
+              next unless (percentage = status&.storage&.dig('pulp_dir', 'percent'))
+
               if percentage[0..2].to_i < 90 && notification_already_exists?(proxy)
                 blueprint.notifications.where(subject: proxy).destroy_all
               elsif update_notifications(proxy).empty? && percentage[0..2].to_i > 90
